Bug in FNR switching - updated throttle voltage ref factor - changed double sin/cos to float sin/cos

+++ Core/Src/FOC_MTPA_FWC_FF.c
@@ -724,13 +724,13 @@ void FOC_MTPA_FWC_FF_step0(void)       /* Sample time: [0.0001s, 0.0s] */
    *  Inport: '<Root>/angle'
    *  Trigonometry: '<S6>/Trigonometric Function1'
    */
-  rtb_TrigonometricFunction_tmp = sin(FOC_MTPA_FWC_FF_U.MtrElcPos);
+  rtb_TrigonometricFunction_tmp = sinf(FOC_MTPA_FWC_FF_U.MtrElcPos);
 
   /* Trigonometry: '<S4>/Trigonometric Function1' incorporates:
    *  Inport: '<Root>/angle'
    *  Trigonometry: '<S6>/Trigonometric Function'
    */
-  rtb_TrigonometricFunction1_tmp = cos(FOC_MTPA_FWC_FF_U.MtrElcPos);
+  rtb_TrigonometricFunction1_tmp = cosf(FOC_MTPA_FWC_FF_U.MtrElcPos);
 
   /* Outputs for Atomic SubSystem: '<S13>/Two inputs CRL' */
   /* Outputs for Atomic SubSystem: '<S14>/Two phase CRL wrap' */
